+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,12 @@
 # babel-preset-dlight
 
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @dlightjs/view-generator@1.0.0-beta.3
+
 ## 1.0.0-beta.8
 
 ### Patch Changes