+++ packages/tools/vite-plugin-dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vite-plugin-dlight
 
+## 1.0.0-beta.9
+
+### Patch Changes
+
+- babel-preset-dlight@1.0.0-beta.9
+
 ## 1.0.0-beta.8
 
 ### Patch Changes